Add livedata & viewmodel (II)
