+++ MoviesExample/app/src/main/java/com/novoda/demo/movies/MoviesViewModelFactory.java
@@ -4,19 +4,21 @@ import android.arch.lifecycle.ViewModel;
 import android.arch.lifecycle.ViewModelProvider;
 import android.support.annotation.NonNull;
 
+import com.novoda.demo.movies.api.MoviesApi;
+
 public class MoviesViewModelFactory implements ViewModelProvider.Factory{
 
-    private MovieService movieService;
+    private MoviesApi moviesApi;
 
-    MoviesViewModelFactory(MovieService movieService) {
-        this.movieService = movieService;
+    MoviesViewModelFactory(MoviesApi moviesApi) {
+        this.moviesApi = moviesApi;
     }
 
     @NonNull
     @Override
     public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
         if (modelClass.isAssignableFrom(MoviesViewModel.class)) {
-            return (T) new MoviesViewModel(movieService);
+            return (T) new MoviesViewModel(moviesApi);
         }
         throw new IllegalArgumentException("Unknown ViewModel class");
     }
