+++ MoviesExample/core/build.gradle
@@ -5,7 +5,6 @@ dependencies {
 
     compile 'com.squareup.retrofit2:retrofit:2.2.0'
     compile 'com.squareup.retrofit2:converter-gson:2.2.0'
-    compile 'com.squareup.retrofit2:adapter-rxjava:2.2.0'
 }
 
 sourceCompatibility = "1.7"
