+++ MoviesExample/core/src/main/java/com/novoda/demo/movies/Dependencies.java
@@ -7,7 +7,6 @@ import java.io.File;
 import okhttp3.Cache;
 import okhttp3.OkHttpClient;
 import retrofit2.Retrofit;
-import retrofit2.adapter.rxjava.RxJavaCallAdapterFactory;
 import retrofit2.converter.gson.GsonConverterFactory;
 
 public class Dependencies {
@@ -34,7 +33,6 @@ public class Dependencies {
                     .client(httpClient)
                     .baseUrl("https://api.themoviedb.org/3/")
                     .addConverterFactory(GsonConverterFactory.create())
-                    .addCallAdapterFactory(RxJavaCallAdapterFactory.create())
                     .build();
     }
